refactor docker image names used in the tests
