+++ .golangci.yaml
@@ -146,8 +146,8 @@ linters-settings:
         alias: slicesutils
       - pkg: github.com/kyma-project/telemetry-manager/internal/utils/test
         alias: testutils
-      - pkg: github.com/kyma-project/telemetry-manager/internal/common
-        alias: commonconsts
+      - pkg: github.com/kyma-project/telemetry-manager/internal/constant
+        alias: constants
   errcheck:
     check-type-assertions: true # Reports type assertions: `a := b.(SomeStruct)`.
     check-blank: true # Report assignment of errors to blank identifier: `num, _ := strconv.Atoi(numStr)`.
