+++ internal/resources/otelcollector/testdata/log-gateway.yaml
@@ -137,7 +137,7 @@ spec:
         - secretRef:
             name: telemetry-log-gateway
             optional: true
-        image: opentelemetry/collector:latest
+        image: opentelemetry/collector:dummy
         livenessProbe:
           httpGet:
             path: /
