+++ internal/resources/otelcollector/testdata/metric-gateway.yaml
@@ -138,7 +138,7 @@ spec:
         - secretRef:
             name: telemetry-metric-gateway
             optional: true
-        image: opentelemetry/collector:latest
+        image: opentelemetry/collector:dummy
         livenessProbe:
           httpGet:
             path: /
