+++ internal/resources/otelcollector/testdata/trace-gateway.yaml
@@ -138,7 +138,7 @@ spec:
         - secretRef:
             name: telemetry-trace-gateway
             optional: true
-        image: opentelemetry/collector:latest
+        image: opentelemetry/collector:dummy
         livenessProbe:
           httpGet:
             path: /
