+++ internal/tools/generate.go
@@ -41,7 +41,7 @@ func main() {
 }
 
 func generateCommonConstants(data map[string]string) {
-	f, err := os.Create("../common/const.go")
+	f, err := os.Create("../constant/images.go")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -49,7 +49,7 @@ func generateCommonConstants(data map[string]string) {
 	_, err = f.WriteString(`// This file is generated by "make generate-image-const".
 // Don't edit, update .env file and run make target
 
-package common
+package constant
 
 `)
 
@@ -105,7 +105,7 @@ package common
 }
 
 func generateTestKitCommonConstants(data map[string]string) {
-	f, err := os.Create("../../test/testkit/const.go")
+	f, err := os.Create("../../test/testkit/images.go")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -126,16 +126,7 @@ package testkit
 		log.Fatal(err)
 	}
 
-	_, err = f.WriteString("\tDefaultOTelCollectorImage = ")
-	if err != nil {
-		log.Fatal(err)
-	}
-	_, err = f.WriteString(fmt.Sprintf("\"%s\"\n", data["DEFAULT_OTEL_COLLECTOR_IMAGE"]))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	_, err = f.WriteString("\tDefaultTelemetryGenImage  = ")
+	_, err = f.WriteString("\tDefaultTelemetryGenImage = ")
 	if err != nil {
 		log.Fatal(err)
 	}
