--- test/testkit/const.go
@@ -1,9 +1,0 @@
-// This file is generated by "make generate-image-const".
-// Don't edit, update .env file and run make target
-
-package testkit
-
-const (
-	DefaultOTelCollectorImage = "europe-docker.pkg.dev/kyma-project/prod/kyma-otel-collector:0.114.0-1.29.0"
-	DefaultTelemetryGenImage  = "ghcr.io/open-telemetry/opentelemetry-collector-contrib/telemetrygen:v0.115.0"
-)
