+++ test/testkit/images.go
@@ -1,0 +1,8 @@
+// This file is generated by "make generate-image-const".
+// Don't edit, update .env file and run make target
+
+package testkit
+
+const (
+	DefaultTelemetryGenImage = "ghcr.io/open-telemetry/opentelemetry-collector-contrib/telemetrygen:v0.115.0"
+)
